Crate storage_proto

This crate provides Protocol Buffers definitions for the services provided by the storage_service crate.

The protocol is documented in Protocol Buffers sources files in the .proto extension and the documentation is not viewable via rustdoc. Refer to the source code to see it.

The content provided in this documentation falls to two categories:

Those automatically generated by grpc-rs:

In [proto::storage] are structs corresponding to our Protocol Buffers messages.

In [proto::storage_grpc] live the GRPC client struct and the service trait which correspond to our Protocol Buffers services.

Structs we wrote manually as helpers to ease the manipulation of the above category of structs. By implementing the TryFrom and From traits, these structs convert from/to the above category of structs in a single method call and in that process data integrity check can be done. These live right in the root module of this crate (this page).

Ihis is provided as a separate crate so that crates that use the storage service via storage-client don't need to depending on the entire storage_service.

Modules

prelude

proto

Structs

GetAccountStateWithProofByVersionRequest
Helper to construct and parse [proto::storage::GetAccountStateWithProofByVersionRequest]

GetAccountStateWithProofByVersionResponse
Helper to construct and parse [proto::storage::GetAccountStateWithProofByVersionResponse]

GetLatestLedgerInfosPerEpochRequest
Helper to construct and parse [proto::storage::GetLatestLedgerInfosPerEpochRequest]

GetLatestLedgerInfosPerEpochResponse
Helper to construct and parse [proto::storage::GetLatestLedgerInfosPerEpochResponse]

GetStartupInfoResponse
Helper to construct and parse [proto::storage::GetStartupInfoResponse]

GetTransactionsRequest
Helper to construct and parse [proto::storage::GetTransactionsRequest]

GetTransactionsResponse
Helper to construct and parse [proto::storage::GetTransactionsResponse]

SaveTransactionsRequest
Helper to construct and parse [proto::storage::SaveTransactionsRequest]

StartupInfo
Helper to construct and parse [proto::storage::StartupInfo]

TreeState