Trait storage_proto::proto::storage::storage_server::Storage

pub trait Storage: Send + Sync + 'static {
    type BackupAccountStateStream: Stream<Item = Result<BackupAccountStateResponse, Status>> + Send + Sync + 'static;
    type BackupTransactionStream: Stream<Item = Result<BackupTransactionResponse, Status>> + Send + Sync + 'static;
    type BackupTransactionInfoStream: Stream<Item = Result<BackupTransactionInfoResponse, Status>> + Send + Sync + 'static;
#[must_use]
    fn save_transactions<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<SaveTransactionsRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<SaveTransactionsResponse>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]
    fn update_to_latest_ledger<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<UpdateToLatestLedgerRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<UpdateToLatestLedgerResponse>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]
    fn get_transactions<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<GetTransactionsRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<GetTransactionsResponse>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]
    fn get_latest_state_root<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<GetLatestStateRootRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<GetLatestStateRootResponse>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]
    fn get_latest_account_state<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<GetLatestAccountStateRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<GetLatestAccountStateResponse>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]
    fn get_account_state_with_proof_by_version<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<GetAccountStateWithProofByVersionRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<GetAccountStateWithProofByVersionResponse>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]
    fn get_startup_info<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<GetStartupInfoRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<GetStartupInfoResponse>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]
    fn get_epoch_change_ledger_infos<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<GetEpochChangeLedgerInfosRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<EpochChangeProof>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]
    fn backup_account_state<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<BackupAccountStateRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<Self::BackupAccountStateStream>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]
    fn get_account_state_range_proof<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<GetAccountStateRangeProofRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<GetAccountStateRangeProofResponse>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]
    fn backup_transaction<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<BackupTransactionRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<Self::BackupTransactionStream>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]
    fn backup_transaction_info<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<BackupTransactionInfoRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<Self::BackupTransactionInfoStream>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
}

Generated trait containing gRPC methods that should be implemented for use with StorageServer.

Associated Types

type BackupAccountStateStream: Stream<Item = Result<BackupAccountStateResponse, Status>> + Send + Sync + 'static

Server streaming response type for the BackupAccountState method.

type BackupTransactionStream: Stream<Item = Result<BackupTransactionResponse, Status>> + Send + Sync + 'static

Server streaming response type for the BackupTransaction method.

type BackupTransactionInfoStream: Stream<Item = Result<BackupTransactionInfoResponse, Status>> + Send + Sync + 'static

Server streaming response type for the BackupTransactionInfo method.

Loading content...
Required methods

#[must_use] fn save_transactions<'life0, 'async_trait>(
&'life0 self,
request: Request<SaveTransactionsRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<SaveTransactionsResponse>, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Persist transactions. Called by Execution when either syncing nodes or committing blocks during normal operation.

#[must_use] fn update_to_latest_ledger<'life0, 'async_trait>(
&'life0 self,
request: Request<UpdateToLatestLedgerRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<UpdateToLatestLedgerResponse>, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Used to get a piece of data and return the proof of it. If the client knows and trusts a ledger info at version v, it should pass v in as the client_known_version and we will return the latest ledger info together with the proof that it derives from v.

#[must_use] fn get_transactions<'life0, 'async_trait>(
&'life0 self,
request: Request<GetTransactionsRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<GetTransactionsResponse>, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

When we receive a request from a peer validator asking a list of transactions for state synchronization, this API can be used to serve the request. Note that the peer should specify a ledger version and all proofs in the response will be relative to this given ledger version.

#[must_use] fn get_latest_state_root<'life0, 'async_trait>(
&'life0 self,
request: Request<GetLatestStateRootRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<GetLatestStateRootResponse>, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

#[must_use] fn get_latest_account_state<'life0, 'async_trait>(
&'life0 self,
request: Request<GetLatestAccountStateRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<GetLatestAccountStateResponse>, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

#[must_use] fn get_account_state_with_proof_by_version<'life0, 'async_trait>(
&'life0 self,
request: Request<GetAccountStateWithProofByVersionRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<GetAccountStateWithProofByVersionResponse>, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

#[must_use] fn get_startup_info<'life0, 'async_trait>(
&'life0 self,
request: Request<GetStartupInfoRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<GetStartupInfoResponse>, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Returns information needed for libra core to start up.

#[must_use] fn get_epoch_change_ledger_infos<'life0, 'async_trait>(
&'life0 self,
request: Request<GetEpochChangeLedgerInfosRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<EpochChangeProof>, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Returns latest ledger infos per epoch.

#[must_use] fn backup_account_state<'life0, 'async_trait>(
&'life0 self,
request: Request<BackupAccountStateRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<Self::BackupAccountStateStream>, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Returns a stream of account states.

#[must_use] fn get_account_state_range_proof<'life0, 'async_trait>(
&'life0 self,
request: Request<GetAccountStateRangeProofRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<GetAccountStateRangeProofResponse>, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

#[must_use] fn backup_transaction<'life0, 'async_trait>(
&'life0 self,
request: Request<BackupTransactionRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<Self::BackupTransactionStream>, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Returns a stream of transactions.

#[must_use] fn backup_transaction_info<'life0, 'async_trait>(
&'life0 self,
request: Request<BackupTransactionInfoRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<Self::BackupTransactionInfoStream>, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Returns a stream of transaction infos.

Loading content...